+++ libvast/vast/system/index.hpp
@@ -242,10 +242,14 @@ pack(flatbuffers::FlatBufferBuilder& builder, const index_state& x);
 /// forwarded to partitions.
 /// @param dir The directory of the index.
 /// @param partition_capacity The maximum number of events per partition.
+/// @param taste_partitions How many lookup partitions to schedule immediately.
+/// @param num_workers The maximum amount of concurrent lookups.
+/// @param meta_index_fp_rate The false positive rate for the meta index.
 /// @pre `partition_capacity > 0
 index_actor::behavior_type
 index(index_actor::stateful_pointer<index_state> self,
       filesystem_actor filesystem, path dir, size_t partition_capacity,
-      size_t in_mem_partitions, size_t taste_partitions, size_t num_workers);
+      size_t in_mem_partitions, size_t taste_partitions, size_t num_workers,
+      double meta_index_fp_rate);
 
 } // namespace vast::system