+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Every entry has a category for which we use the following visual abbreviations:
 
 ## Unreleased
 
+- 🎁 A new type-level synopsis structure in the meta-index now massively
+  speeds up string queries with very few results.
+  [#1214](https://github.com/tenzir/vast/pull/1214)
+
 - ⚡️ The build configuration of VAST received a major overhaul. Inclusion of
   libvast in other procects via `add_subdirectory(path/to/vast)` is now easily
   possible. The names of all build options were aligned, and the new build